# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Chat_API",
      "contact": {},
      "version": "1.0"
    },
    "paths": {
...
      "/api/v1/users/login": {
        
        "post": {
          "tags": [
            "user"
          ],
          "summary": "login user",
          "operationId": "loginuser",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/loginuserrequest"
                    },
                    {
                      "example": {
                        "email": "maria-2828@hotmail.com",
                        "password": "Academlo123%"
                      }
                    }
                  ]
                },
                "example": {
                  "email": "maria-2828@hotmail.com",
                  "password": "Academlo123%"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "loginuserrequest": {
          "title": "loginuserrequest",
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "email": "evamaria@hotmail.com",
            "password": "Academlo123%"
          }
        },
...
      },
      "securitySchemes": {
        "bearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
...
  }